Add user service

+++ server.js
@@ -4,6 +4,8 @@ import { bugService } from './services/bug.service.js'
 import cookieParer from 'cookie-parser'
 import path from 'path'
 import { log } from 'console'
+import { userService } from './services/user.service.js'
+import { authService } from './services/auth.service.js'
 const app = express()
 
 app.use(express.static('public'))
@@ -16,7 +18,7 @@ app.listen(3031, () => {
 })
 //LIST BUG
 app.get('/api/bug', (req, res) => {
-  const queryOptions=preseQueryParams(req.query)
+  const queryOptions = preseQueryParams(req.query)
   bugService.query(queryOptions).then(bugs => res.send(bugs))
     .catch(err => {
       res.status(500).send('Cannot load bugs')
@@ -24,23 +26,26 @@ app.get('/api/bug', (req, res) => {
 
 })
 
-function preseQueryParams(queryParams){
-  const filterBy={
-    txt:queryParams.txt||'',
-    minSeverity:+queryParams.minSeverity||0,
+function preseQueryParams(queryParams) {
+  const filterBy = {
+    txt: queryParams.txt || '',
+    minSeverity: +queryParams.minSeverity || 0,
     pageIdx: queryParams.pageIdx
-   }
-   const sortBy={
-     sortTitle:queryParams.sortField||'',
-     sortSeverity:queryParams.sortField||'',
-     sortDir:+queryParams.sortField||1
- 
-   }
- return {sortBy,filterBy}
+  }
+  const sortBy = {
+    sortTitle: queryParams.sortField || '',
+    sortSeverity: queryParams.sortField || '',
+    sortDir: +queryParams.sortField || 1
+
+  }
+  return { sortBy, filterBy }
 }
 
 //CREATE BUG
 app.post('/api/bug/', (req, res) => {
+  const loggedinUser=authService.validateToken(req.cookies.loginToken)
+  if(!loggedinUser) return res.status(401).send(`Can't remove car`)
+
   const bugToSave = req.body
   bugService.save(bugToSave).then(bug => res.send(bug))
     .catch(err => {
@@ -52,6 +57,9 @@ app.post('/api/bug/', (req, res) => {
 })
 //UPDATE BUG
 app.put('/api/bug/:bugId', (req, res) => {
+  const loggedinUser=authService.validateToken(req.cookies.loginToken)
+  if(!loggedinUser) return res.status(401).send(`Can't remove car`)
+
   const bugToSave = req.body
   bugService.save(bugToSave).then(bug => res.send(bug))
     .catch(err => {
@@ -79,6 +87,8 @@ app.get('/api/bug/:bugId', (req, res) => {
 
 //REMOVE BUG
 app.delete('/api/bug/:bugId', (req, res) => {
+  const loggedinUser=authService.validateToken(req.cookies.loginToken)
+  if(!loggedinUser) return res.status(401).send(`Can't remove car`)
   const { bugId } = req.params
   bugService.remove(bugId)
     .then(() => res.send('bug Removed'))
@@ -90,3 +100,37 @@ app.delete('/api/bug/:bugId', (req, res) => {
 app.get('/**', (req, res) => {
   res.sendFile(path.resolve('public/index.html'))
 })
+
+//// User API
+app.post('/api/auth/signup', (req, res) => {
+  const credentials = req.body
+  userService.add(credentials).then(
+    user => {
+      if (user) {
+        const loginToken = authService.getLoginToken(user)
+        res.cookie('loginToken', loginToken)
+        res.send(user)
+      } else {
+        res.status(400).send('Cannot signup')
+      }
+    }
+  ).catch(err => res.status(400).send('Username taken.'))
+})
+
+app.post('/api/auth/login', (req, res) => {
+  const credentials = req.body
+
+  authService.checkLogin(credentials)
+      .then(user => {
+          const loginToken = authService.getLoginToken(user)
+          res.cookie('loginToken', loginToken)
+          res.send(user)
+      })
+      .catch(() => res.status(404).send('Invalid Credentials'))
+})
+
+app.post('/api/auth/logout', (req, res) => {
+  res.clearCookie('loginToken')
+  res.send('logged-out!')
+})
+
